requirements.txt
aiogram==2.18
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.9.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
logger==1.4
loguru==0.6.0
multidict==6.0.2
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pydantic==1.9.0
pyparsing==3.0.7
pytest==6.2.5
pytest-asyncio==0.17.2
python-dotenv==0.19.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
rsa==4.8
six==1.16.0
Telethon==1.24.0
toml==0.10.2
tornado==6.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
win32-setctime==1.1.0
yarl==1.7.2
